refactor: Move credentials into internal
