+++ src/main/groovy/com/deploygate/gradle/plugins/dsl/DeployGateExtension.groovy
@@ -1,7 +1,7 @@
 package com.deploygate.gradle.plugins.dsl
 
 import com.deploygate.gradle.plugins.DeployGatePlugin
-import com.deploygate.gradle.plugins.credentials.CliCredentialStore
+import com.deploygate.gradle.plugins.internal.credentials.CliCredentialStore
 import com.deploygate.gradle.plugins.dsl.syntax.ExtensionSyntax
 import com.deploygate.gradle.plugins.internal.annotation.DeployGateInternal
 import com.deploygate.gradle.plugins.internal.http.HttpClient
