+++ src/main/groovy/com/deploygate/gradle/plugins/tasks/LogoutTask.groovy
@@ -1,6 +1,6 @@
 package com.deploygate.gradle.plugins.tasks
 
-import com.deploygate.gradle.plugins.credentials.CliCredentialStore
+import com.deploygate.gradle.plugins.internal.credentials.CliCredentialStore
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.TaskAction
