+++ src/test/groovy/com/deploygate/gradle/plugins/tasks/LoginTaskSpec.groovy
@@ -1,6 +1,6 @@
 package com.deploygate.gradle.plugins.tasks
 
-import com.deploygate.gradle.plugins.credentials.CliCredentialStore
+import com.deploygate.gradle.plugins.internal.credentials.CliCredentialStore
 import com.deploygate.gradle.plugins.dsl.DeployGateExtension
 import com.deploygate.gradle.plugins.dsl.NamedDeployment
 import org.gradle.api.NamedDomainObjectContainer
